Updated DSConverter to properly handle WMV
When you try to encode a WMV file with mencode it tries to use a
frame rate of 1000fps (!) as you can imagine this takes forever.
It also means that the resulting file cannot be played in anything.

I added a check to see what the framerate is using mediainfo and then
use that to pass into mencoder.