Fixed bug in RELSEXT where PID was prepended
When working with relsext and passing a string into the function
the library was calling the relsint function which prepended the
pid to the string, so the result ended up looking like this:
PID/PID this is correct when a DSID is passed in as you get
PID/DSID, but incorrect for RELSEXT.

I introduced this when I was optimizing the functions. I sepereated
out the serialization functions for RELSINT and RELSEXT again and thing
should be back to normal now.