grumble .. inconsistant data .. grumble .. tests
It turns out that records that are migrated from fedora 2.x happen to
use a description tag instead of a Description tag. What a pain in the
@*#&. So I updated the fedora relationships library to only create the
Description tag, but deal with the description tag if it exists. If
it does exist we don't change it we just use it.

I also ... dun dun dun ... ADDED TESTS. OMG. I rearranged the library
to make it more testable, then added some tests to the tests folder.
They are sparse right now, but they gave me confidence that both
Description and description work.