Updated DSConverter.py with better checks for PDF
We used to check the return value of PDF2SWF and assume that if it
returned 0, then the conversion was successful. This is not always
the case. If the PDF has security turned on PDF2SWF will output:
FATAL   PDF disallows copying and return 0. This was causing odd
error messages in the logs.

Added a check to make sure the file exists. We also now log whatever
PDF2SWF outputs to STDOUT if the file doesn't exist so that we have a
reasonable log message.